Move deserialize provider into Symfony namespace
